requirements.txt
numpy==1.19.5
matplotlib==3.3.4
Pillow==8.1.2
streamlit==0.78.0
torch==1.8.0
torchvision==0.9.0
